dynamic_ogg.c: fixed ov_time_seek() signature for libvorbisidec.
reported by Vitaly Novichkov.

+++ dynamic_ogg.c
@@ -78,7 +78,7 @@ int Mix_InitOgg()
 		}
 		vorbis.ov_time_seek =
 #ifdef OGG_USE_TREMOR
-			(long (*)(OggVorbis_File *,ogg_int64_t))
+			(int (*)(OggVorbis_File *,ogg_int64_t))
 #else
 			(int (*)(OggVorbis_File *,double))
 #endif
